math fixes

+++ src/Traits/DecodableTrait.php
@@ -31,18 +31,18 @@ trait DecodableTrait
     
         $i = 2;
         for (; $i < count($decoded);) {
-            $deltaLat = (float) bcdiv(self::toSigned($decoded[$i]), $factorDegree, 7);
-            $deltaLng = (float) bcdiv(self::toSigned($decoded[$i + 1]), $factorDegree, 7);
-            $lastLat = (float) bcadd($lastLat, $deltaLat, 7);
-            $lastLng = (float) bcadd($lastLng, $deltaLng, 7);
+            $deltaLat = bcdiv(self::toSigned($decoded[$i]), $factorDegree, 7);
+            $deltaLng = bcdiv(self::toSigned($decoded[$i + 1]), $factorDegree, 7);
+            $lastLat = bcadd($lastLat, $deltaLat, 7);
+            $lastLng = bcadd($lastLng, $deltaLng, 7);
 
             if ($thirdDim) {
                 $deltaZ = self::toSigned($decoded[$i + 2]) / $factorZ;
-                $lastZ = (float) bcadd($lastZ, $deltaZ, 7);
-                $res[] = [$lastLat, $lastLng, $lastZ];
+                $lastZ = bcadd($lastZ, $deltaZ, 7);
+                $res[] = [(float) $lastLat, (float) $lastLng, (float) $lastZ];
                 $i += 3;
             } else {
-                $res[] = [$lastLat, $lastLng];
+                $res[] = [(float) $lastLat, (float) $lastLng];
                 $i += 2;
             }
         }
